Use intmax_t and %jd for time_t, don't assume it's long.
(cherry picked from commit 2608cfc26abd0179d4c08772f91199bb08ced15b)

+++ libpromises/evalfunction.c
@@ -4974,7 +4974,7 @@ static FnCallResult FnCallOn(ARG_UNUSED EvalContext *ctx, ARG_UNUSED const Polic
         Log(LOG_LEVEL_INFO, "Illegal time value");
     }
 
-    return FnReturnF("%ld", cftime);
+    return FnReturnF("%jd", (intmax_t) cftime);
 }
 
 /*********************************************************************/
@@ -5049,7 +5049,7 @@ static FnCallResult FnCallAgoDate(ARG_UNUSED EvalContext *ctx, ARG_UNUSED const 
         return FnReturn("0");
     }
 
-    return FnReturnF("%ld", cftime);
+    return FnReturnF("%jd", (intmax_t) cftime);
 }
 
 /*********************************************************************/
@@ -5067,7 +5067,7 @@ static FnCallResult FnCallAccumulatedDate(ARG_UNUSED EvalContext *ctx, ARG_UNUSE
     cftime += tmv.tm_mon * 30 * 24 * 3600;
     cftime += (tmv.tm_year + 1900) * 365 * 24 * 3600;
 
-    return FnReturnF("%ld", cftime);
+    return FnReturnF("%jd", (intmax_t) cftime);
 }
 
 /*********************************************************************/
@@ -5084,7 +5084,7 @@ static FnCallResult FnCallNot(EvalContext *ctx,
 
 static FnCallResult FnCallNow(ARG_UNUSED EvalContext *ctx, ARG_UNUSED const Policy *policy, ARG_UNUSED const FnCall *fp, ARG_UNUSED const Rlist *finalargs)
 {
-    return FnReturnF("%ld", (long)CFSTARTTIME);
+    return FnReturnF("%jd", (intmax_t)CFSTARTTIME);
 }
 
 /*********************************************************************/
